Added the bare basics of code generation. It now prints hello world each time.
The .j file is generated in the same folder as the .lol file. This can be changed if it's not wanted.

Having trouble linking the PascalRTL.jar to the project, I'm not sure how to get jasmin to be able to use those methods etc.

Updated gitignore to ignore generated .j files.

Goodnight!