Minor changes to several things. Break statements now have a stack, for when there are loops in loops or loops in methods.
Added test prog for if statements. It currently crashes while parsing, I'll work on if logic in the morning to fix that.

I think IT can be handled by putting it in a local variable slot, slot 0 or something, just need to set it each time a LOL_STD_OP returns. Also modify assignment and variable declaration operations to work with this implementation. I'll see if I can get that to work after if's, or during ifs since ifs kinda rely on IT